Module Oasis.Builtins

Common functions you need everywhere

val abs : int -> int

Absolute value of integers

val absf : float -> float

Absolute value of floats

val pow : int -> int -> float

Power of ints

val powf : float -> float -> float

Power of floats

val divmod : int -> int -> int * int

Get pair of dividend and modulo of two integers

val divmodf : float -> float -> float * float

Get pair of dividend and modulo of two floats https://stackoverflow.com/a/37977740

val round_dfrac : int -> float -> float

module Complex : sig ... end

Complex numbers

type complex = Complex.t

val base_number : ?first:bool -> string -> int -> int -> string

Get number in bases up to 36

val bin : int -> string

Binary string of number

val oct : int -> string

Octal string of number

val hex : int -> string

Hexadecimal string of number

val chr : int -> string

Get string represented by a unicode code point

val ord : string -> int

Get unicode code point of a string

val format : ('a, unit, string) Stdlib.format -> 'a

Get the formatted string for a given object

module Map : sig ... end

Maps/Dictionaries

type ('a, 'b) map = ('a, 'b) Map.t

val all : bool list -> bool

All items in list are true

val any : bool list -> bool

Any item in list is true

val enumerate : ?start:int -> 'a list -> (int * 'a) list

Get index-annotated list of items of a list

val filter : ('a -> bool) -> 'a list -> 'a list

Filter a list by a given selection function

val sort : ('a -> 'a -> int) -> 'a list -> 'a list

Sort a list by a given comparison function

val map : ('a -> 'b) -> 'a list -> 'b list

Map a function onto a list

val get_superlative : 'a list -> ('a -> 'a -> float) -> 'a option

Get most <metric> item in list by a given comparison function

val max_int : int list -> int option

Get largest int in list

val max_float : float list -> float option

Get largest float in list

val min_int : int list -> int option

Get smallest int in list

val min_float : float list -> float option

Get smallest float in list

val max_complex : complex list -> complex option

Get smallest complex in list

val range : ?start:int -> ?step:int -> int -> int list

Get range a..b)

val reversed : 'a list -> 'a list

Reverse a list

val unique : 'a list -> 'a list

Get unique items of list

val slice : 'a list -> ?step:int -> int -> int -> 'a list

Slice a list

val sum : int list -> int

Sum an int list

val sumf : float list -> float

Sum a float list

val sumc : complex list -> Complex.t

Sum a complex list

val zip : 'a list -> 'b list -> ('a * 'b) list

Combine pairs of lists into lists of pairs

val stdinput : string -> string

Get input from stdin

val open_in_file : string -> int -> string -> Stdlib.out_channel

Open an output file