Module Collections.Counter

Count items in a list

include module type of struct include Builtins.Dict end

Dictionaries

val (.?[]) : ('a, 'b) Builtins.Dict.t -> 'a -> 'b

Get operator

val empty : 'b -> ('a, 'b) Builtins.Dict.t

Empty dictionary with default value

exception KeyError of string

No matching key error

val err_empty : ?string_of_key:('a -> string) option -> unit -> ('a, 'b) Builtins.Dict.t

Empty dictionary without default value, raises error

val safe_get : ('a, 'b) Builtins.Dict.t -> 'a -> 'b option

Safe get, catches KeyErrors

val len : ('a, 'b) Builtins.Dict.t -> int

Number of stored items

val key_in : ('a, 'b) Builtins.Dict.t -> 'a -> bool

Whether a key exists in the dictionary

val val_in : ('a, 'b) Builtins.Dict.t -> 'b -> bool

Whether a value exists in the dictionary

val update : ('a, 'b) Builtins.Dict.t -> 'a -> 'b -> ('a, 'b) Builtins.Dict.t

Add to dictionary

val (.?[]<-) : ('a, 'b) Builtins.Dict.t -> 'a -> 'b -> ('a, 'b) Builtins.Dict.t

Set operator

val from_keys : ('a, 'b) Builtins.Dict.t -> 'b -> ('a, 'b) Builtins.Dict.t

Get an empty dictionary with the same keys as another

val keys : ('a, 'b) Builtins.Dict.t -> 'a list

Public access to dictionary keys

val values : ('a, 'b) Builtins.Dict.t -> 'b list

Public access to dictionary values

val items : ('a, 'b) Builtins.Dict.t -> ('a * 'b) list

Key-value pairs of dictionary

val merge : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t

Add key-value pairs from another dictionary

val subset : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> bool

val (<?) : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> bool

val (>?) : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> bool

val (=?) : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> bool

Equality over keys and values (order doesn't matter)

val (<>?) : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> bool

val (<=?) : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> bool

val (>=?) : ('a, 'b) Builtins.Dict.t -> ('a, 'b) Builtins.Dict.t -> bool

type 'a t = ('a, int) Builtins.dict

val counter : 'a list -> 'a t

Get counter object for a list

val elements : 'a t -> 'a list

Get elements of original list, grouped and in order of retrieval

val most_common : ?n:int -> 'a t -> ('a * int) list

Get the most common elements of original list, annotated with their occurrence numbers

parameter x
The counter to look in

parameter n
The number of most common elements to retrieve, defaults to -1 (all)

returns
A list of pairs of items and their occurrence numbers

val subtract : 'a t -> 'a t -> 'a t

Elements occurrence values from one counter are subtracted from another, with the value for missing element occurrences being 0

parameter a
The counter to subtract from

parameter b
The counter being subtracted

returns
The remainder counter

val total : 'a t -> int

The total value of the counts of a counter