Module Collections.Heap

Binary heap implementation

Parameters

module OT : Stdlib.Map.OrderedType

Signature

include module type of struct include Stdlib.List end

val length : 'a list -> int

val compare_lengths : 'a list -> 'b list -> int

val compare_length_with : 'a list -> int -> int

val is_empty : 'a list -> bool

val cons : 'a -> 'a list -> 'a list

val hd : 'a list -> 'a

val tl : 'a list -> 'a list

val nth : 'a list -> int -> 'a

val nth_opt : 'a list -> int -> 'a option

val rev : 'a list -> 'a list

val init : int -> (int -> 'a) -> 'a list

val append : 'a list -> 'a list -> 'a list

val rev_append : 'a list -> 'a list -> 'a list

val concat : 'a list list -> 'a list

val flatten : 'a list list -> 'a list

val equal : ('a -> 'a -> bool) -> 'a list -> 'a list -> bool

val compare : ('a -> 'a -> int) -> 'a list -> 'a list -> int

val iter : ('a -> unit) -> 'a list -> unit

val iteri : (int -> 'a -> unit) -> 'a list -> unit

val map : ('a -> 'b) -> 'a list -> 'b list

val mapi : (int -> 'a -> 'b) -> 'a list -> 'b list

val rev_map : ('a -> 'b) -> 'a list -> 'b list

val filter_map : ('a -> 'b option) -> 'a list -> 'b list

val concat_map : ('a -> 'b list) -> 'a list -> 'b list

val fold_left_map : ('acc -> 'a -> 'acc * 'b) -> 'acc -> 'a list -> 'acc * 'b list

val fold_left : ('acc -> 'a -> 'acc) -> 'acc -> 'a list -> 'acc

val fold_right : ('a -> 'acc -> 'acc) -> 'a list -> 'acc -> 'acc

val iter2 : ('a -> 'b -> unit) -> 'a list -> 'b list -> unit

val map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val rev_map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val fold_left2 : ('acc -> 'a -> 'b -> 'acc) -> 'acc -> 'a list -> 'b list -> 'acc

val fold_right2 : ('a -> 'b -> 'acc -> 'acc) -> 'a list -> 'b list -> 'acc -> 'acc

val for_all : ('a -> bool) -> 'a list -> bool

val exists : ('a -> bool) -> 'a list -> bool

val for_all2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val exists2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val mem : 'a -> 'a list -> bool

val memq : 'a -> 'a list -> bool

val find : ('a -> bool) -> 'a list -> 'a

val find_opt : ('a -> bool) -> 'a list -> 'a option

val find_index : ('a -> bool) -> 'a list -> int option

val find_map : ('a -> 'b option) -> 'a list -> 'b option

val find_mapi : (int -> 'a -> 'b option) -> 'a list -> 'b option

val filter : ('a -> bool) -> 'a list -> 'a list

val find_all : ('a -> bool) -> 'a list -> 'a list

val filteri : (int -> 'a -> bool) -> 'a list -> 'a list

val take : int -> 'a list -> 'a list

val drop : int -> 'a list -> 'a list

val take_while : ('a -> bool) -> 'a list -> 'a list

val drop_while : ('a -> bool) -> 'a list -> 'a list

val partition : ('a -> bool) -> 'a list -> 'a list * 'a list

val partition_map : ('a -> ('b, 'c) Stdlib.Either.t) -> 'a list -> 'b list * 'c list

val assoc : 'a -> ('a * 'b) list -> 'b

val assoc_opt : 'a -> ('a * 'b) list -> 'b option

val assq : 'a -> ('a * 'b) list -> 'b

val assq_opt : 'a -> ('a * 'b) list -> 'b option

val mem_assoc : 'a -> ('a * 'b) list -> bool

val mem_assq : 'a -> ('a * 'b) list -> bool

val remove_assoc : 'a -> ('a * 'b) list -> ('a * 'b) list

val remove_assq : 'a -> ('a * 'b) list -> ('a * 'b) list

val split : ('a * 'b) list -> 'a list * 'b list

val combine : 'a list -> 'b list -> ('a * 'b) list

val sort : ('a -> 'a -> int) -> 'a list -> 'a list

val stable_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val fast_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val sort_uniq : ('a -> 'a -> int) -> 'a list -> 'a list

val merge : ('a -> 'a -> int) -> 'a list -> 'a list -> 'a list

val to_seq : 'a list -> 'a Stdlib.Seq.t

val of_seq : 'a Stdlib.Seq.t -> 'a list

type x = OT.t

type t = x option list

val _oob_option : 'a option -> 'a

val parent : t -> int -> int option

Get the index of the parent of the element at index i in the heap. Returns None if i is the root or invalid.

parameter h
The heap

parameter i
The index of the element

returns
Some index of the parent or None if no parent exists

val lefti : t -> int -> int option

Get the index of the left child of the element at index i, if it exists.

parameter h
The heap

parameter i
The index of the element

returns
Some value if left child exists, else None

val righti : t -> int -> int option

Get the index of the right child of the element at index i, if it exists.

parameter h
The heap

parameter i
The index of the element

returns
Some index if right child exists, else None

val max_heapify : t -> int -> int -> t

Recursively enforce the max-heap property from a given index.

parameter a
The heap (list) to adjust

parameter n
Effective length of a to process

parameter i
The index to enforce heap property from

returns
The adjusted heap

val heapify : x list -> t

Convert a list into a max-heap.

parameter a
The list to heapify

returns
The max-heap represented as a list

val heap_sort : x list -> x list

Sort a list

val heap_increase_key : t -> int -> x -> t

Increase the value of an item in the heap

parameter h
The heap to update

parameter i
The index of the heap to update

parameter k
The new value of h[i]

returns
The updated heap

raises Invalid_Argument
if k < nth h i

raises Failure
if length h <= i

val extract_max : t -> (x * t) option

Extract the maximum element from a max-heap and return the updated heap.

parameter h
The heap

returns
Some (max_element, new_heap) if the heap is non-empty, None otherwise

val heap_insert : t -> x -> t

Insert a new key into the max-heap while maintaining the max-heap property.

parameter h
The current heap

parameter k
The key to insert

returns
The updated heap with the new key inserted and the max-heap property restored