Module Oasis.String

Common string operations

val ascii_lowercase : string

All lowercase ascii letters

val ascii_uppercase : string

All uppercase ascii letters

val ascii_letters : string

All ascii letters

val digits : string

All decimal digit characters

val hexdigits : string

All hexadecimal digit characters

val octdigits : string

All octal digit characters

val punctuation : string

All puncutation characters

val whitespace : string

All whitespace characters

val printable : string

All printable characters

val char_is_printable : char -> bool

Check if a character is printable

val string_is_printable : string -> bool

Check if a string is printable

val template : string -> (string -> string) -> string

Given a template string and a map, fill in the template string. Map keys are denoted by $identifier, and dollar signs are escaped like $$

val capitalize : string -> string

Capitalize first letter of string, lowercase the rest

val center : ?fill_char:char -> string -> int -> string

Center a string with spaces given a target width

val count : string -> string -> int

Count non-overlapping occurances of a substring

val ends_with : string -> string -> bool

endswith base suffix checks if the base string ends with the suffix string

val expand_tabs : ?tab_size:int -> string -> string

Replace tab characters with a specified number of spaces

val find : string -> string -> int

Find first occurance of substring in string